change and support different sentence transformer models

+++ create_embeddings.py
@@ -1,4 +1,5 @@
 import os
+import sys
 
 import numpy as np
 import pandas as pd
@@ -9,14 +10,20 @@ from app.config import settings
 
 def create_embeddings() -> None:
     if os.path.isfile(settings.EMBEDDINGS_FILE):
-        print(f"File already exists at {settings.EMBEDDINGS_FILE}")
-        return
+        print(f"File already exists at {settings.EMBEDDINGS_FILE}", file=sys.stderr)
+        return None
+
+    if not os.path.isfile(settings.CAPTIONS_FILE):
+        print(
+            f"Could not find captions file at {settings.CAPTIONS_FILE}", file=sys.stderr
+        )
+        return None
 
     df = pd.read_csv(settings.CAPTIONS_FILE, delimiter=";")
-    model = SentenceTransformer("bert-base-nli-mean-tokens")
+    model = SentenceTransformer(settings.SENTENCE_TRANSFORMER_MODEL)
 
     print("Creating embeddings")
-    embeddings = model.encode(list(df.top_aggregate_caption))
+    embeddings = model.encode(list(df.top_aggregate_caption), show_progress_bar=True)
 
     print(f"Saving Embeddings to {settings.EMBEDDINGS_FILE}")
     np.save(settings.EMBEDDINGS_FILE, embeddings)
